requirements.txt
python-telegram-bot==13.1
beautifulsoup4==4.9.3
requests==2.24.0
